requirements.txt
google-genai==1.69.0
python-dotenv==1.2.2
pandas==2.3.3
matplotlib==3.10.8
plotly==6.6.0
scikit-learn==1.8.0
streamlit==1.55.0
